Corrigo dtos

+++ src/main/java/com/grupo14/turnos/dto/EspecificacionDTO.java
@@ -5,11 +5,11 @@ import com.grupo14.turnos.modelo.Rubro;
 import jakarta.validation.constraints.NotNull;
 
 public record EspecificacionDTO(
-	    Integer id,
-	    @NotNull Long servicioId,
-	    @NotNull Rubro rubro,
-	    @NotNull String detalles,
-	    @NotNull Integer direccionId,
-	    String servicioNombre,
-	    String direccionTexto
-	) {}
+    Long id,                      
+    @NotNull Long servicioId,
+    @NotNull Rubro rubro,
+    @NotNull String detalles,
+    @NotNull Long direccionId,    
+    String servicioNombre,
+    String direccionTexto
+) {}
